Begin cleanup of ActionForm
